requirements.txt
jsonschema
requests
psycopg2
pytest
pytest-dotenv
pyyaml